+++ example/shieldwall/endcap.py
@@ -3,10 +3,10 @@ from cqportal.shieldwall import EndCap
 
 cap_bp =  EndCap()
 cap_bp.length = 15
-cap_bp.height = 50
-cap_bp.base_height = 5
-cap_bp.side_height = 1
-cap_bp.middle_width_inset = -6
+#cap_bp.height = 50
+#cap_bp.base_height = 5.6
+#cap_bp.side_height = 1
+#cap_bp.middle_width_inset = -6
 
 #cap_bp.greeble_bp.operation = 'fillet'
 cap_bp.greeble_bp.operation = 'chamfer'